KH3 mods and workspace notion
We don't need to run the game in its install directory. In fact, it's better not to, to avoid making random cache files.
DLLs still need to be present there (symlinked), but settings files like for luabackend, panacea, and refined go in the working directory (also symlinked).
This keeps our install folders even cleaner.